Swap function from ATFTable to ATFQuery
