+++ inc/Engine/Media/AboveTheFold/Database/Queries/AboveTheFold.php
@@ -106,24 +106,35 @@ class AboveTheFold extends AbstractQuery {
 	}
 
 	/**
-	 * Deletes rows from the 'above_the_fold' table that have a 'failed' status and have not been accessed since a specified date.
+	 * Delete all rows which were not accessed in the last month.
 	 *
-	 * @return int|false The number of rows affected if the query is successful, or false on failure.
+	 * @return bool|int
 	 */
-	public function delete_failed_and_not_accessed_rows() {
-		// Get the database object.
+	public function delete_old_rows() {
+		// Get the database interface.
 		$db = $this->get_db();
 
-		// If no database interface is available, return false.
+		// Bail if no database interface is available.
 		if ( empty( $db ) ) {
 			return false;
 		}
 
-		// Prepare the SQL query. The '%s' placeholder will be replaced by the value of $date.
-		// The query will delete rows from the 'above_the_fold' table where 'status' is 'failed' and 'last_accessed' is less than or equal to $date.
-		$query = $db->prepare( "DELETE FROM {$this->table_name} WHERE status = 'failed' AND last_accessed <= %s", $date );
+		/**
+		 * Filters the interval (in months) to determine when an Above The Fold (ATF) entry is considered 'old'.
+		 * Old ATF entries are eligible for deletion. By default, an ATF entry is considered old if it hasn't been accessed in the last month.
+		 *
+		 * @param int $delete_interval The interval in months after which an ATF entry is considered old. Default is 1 month.
+		 */
+		$delete_interval = (int) apply_filters( 'rocket_atf_cleanup_interval', 1 );
 
-		// Execute the query and return the result. If the query is successful, the number of affected rows will be returned. If the query fails, false will be returned.
-		return $db->query( $query );
+		if ( $delete_interval <= 0 ) {
+			return false;
+		}
+
+		$prefixed_table_name = $this->apply_prefix( $this->table_name );
+		$query               = "DELETE FROM `$prefixed_table_name` WHERE status = 'failed' OR `last_accessed` <= date_sub(now(), interval $delete_interval month)";
+		$rows_affected       = $db->query( $query );
+
+		return $rows_affected;
 	}
 }
