+++ inc/Engine/Media/AboveTheFold/Database/Tables/AboveTheFold.php
@@ -74,40 +74,4 @@ class AboveTheFold extends AbstractTable {
 
 		return $this->truncate();
 	}
-
-	/**
-	 * Delete all rows which were not accessed in the last month.
-	 *
-	 * @return bool|int
-	 */
-	public function delete_old_rows() {
-		if ( ! $this->exists() ) {
-			return false;
-		}
-		// Get the database interface.
-		$db = $this->get_db();
-
-		// Bail if no database interface is available.
-		if ( empty( $db ) ) {
-			return false;
-		}
-
-		/**
-		 * Filters the interval (in months) to determine when an Above The Fold (ATF) entry is considered 'old'.
-		 * Old ATF entries are eligible for deletion. By default, an ATF entry is considered old if it hasn't been accessed in the last month.
-		 *
-		 * @param int $delete_interval The interval in months after which an ATF entry is considered old. Default is 1 month.
-		 */
-		$delete_interval = (int) apply_filters( 'rocket_atf_cleanup_interval', 1 );
-
-		if ( $delete_interval <= 0 ) {
-			return false;
-		}
-
-		$prefixed_table_name = $this->apply_prefix( $this->table_name );
-		$query               = "DELETE FROM `$prefixed_table_name` WHERE status = 'failed' OR `last_accessed` <= date_sub(now(), interval $delete_interval month)";
-		$rows_affected       = $db->query( $query );
-
-		return $rows_affected;
-	}
 }
